Refactor control-panel div from App into a WorkspaceControlPanel component.
Rationale being that issues like #1586 and #1587 will be able to build on top of this and not have to all work in the top-level App component.
