+++ src/styles/index.scss
@@ -4,7 +4,7 @@ body {
 }
 
 .mirador {
-  &-control-panel {
+  &-workspace-control-panel {
     position: absolute;
     box-sizing: border-box;
     padding: 15px;
